requirements.txt
datetime
ntplib
pypiwin32
python-dateutil